server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai-whisper>=20231117
python-multipart==0.0.17
